+++ Magic/CHANGELOG.md
@@ -22,6 +22,7 @@
  - Fix EffectLib's ExplodeEffect and BigBangEffect
  - Fix being able to put `unstashable` wands in chests if they were in your main hand
  - Fix an error that could occor if a spell kills its own caster inside of an Asyncronous action
+ - Update BattleArena integration
 
 ## Changes